+++ axum/src/extract/mod.rs
@@ -39,11 +39,8 @@ pub use crate::Json;
 pub use crate::Extension;
 
 #[cfg(feature = "form")]
-mod form;
-
-#[cfg(feature = "form")]
-#[doc(inline)]
-pub use self::form::Form;
+#[doc(no_inline)]
+pub use crate::form::Form;
 
 #[cfg(feature = "matched-path")]
 mod matched_path;